Basic IK algorithm & more controller features + refactoring
Implemented a basic IK algorithm. While it works, I fully expect problems to appear if I don't limit the possible angles for the servos. Self intersection is not accounted for.

Controller class was refactored to be more concise + added a factor that accounts for the strength with which the joystick is pushed.